getversion: fix handling of detached commits
The inline documentation of getversion claims to use git describe for
detached heads. Presumably git's output for detached heads changed from
something that starts with "(no" to something that starts with
"(detached" now. Contrary to the documentation, detached branches were
handled like feature branches with a "(" in their name which produces a
version string such as:

0.0.6-(detached-2013-10-24-c12-gabcdef

The opening brace breaks autoreconf like this:

/usr/bin/m4:configure.ac:11: ERROR: end of file in argument list