Read enough bytes to avoid output truncation
With gcc-7 the option -Wformat-truncation can be used to warn about
possible output truncation when using snprintf() and vsnprintf().

The line in question here creates an output of max 269 bytes:  maximum
size of e->d_name (afaik 255 byte) plus the 13 bytes of the string "/proc//status" but
read only 256 byte into buf.

So increase the size of char buf[] and increase the number of bytes
read into buf silences the warning.

This fixes Debian #871108.

See
https://gcc.gnu.org/onlinedocs/gcc-7.2.0/gcc/Warning-Options.html#Warning-Options