Use file RELEASE as source for the version during build
That commit allows to build from tarball releases. This also fixes
Debian bug #845426 but without adding Debian specifics as proposed by
the reporter.

Older tarball releases of munin-c does not contained the file RELEASE even if
'getversion' in the past supported it (most likely because getversion
was copied from the Munin repository). So the release process needs
an additional step of adding the RELEASE file.