Reformat source to match kernel coding style
While reviewing a pull request i noticed a different indention style and
tried to fix that in an automated way. But i wasn't able to describe the
rules for the actual coding style. So i looked up what the kernel devs
do and found a useable indent command line here:

https://www.kernel.org/doc/html/v4.20/process/coding-style.html

If we agree in using the kernel coding style the following indent
commandline can be use to format the source accordingly:

  indent -kr -i8

Signed-off-by: Matthias Schmitz <matthias@sigxcpu.org>