t: cleanup the testing suite
We now test the testing suite properly, with "make check". Also having
the listplugins isn't used anymore as it is not part of the program
install, but configuration of an eventual package

Also, we don't want to require any execution for non test make targets,
as it would cause issues with cross compilation.

If cross-testing needs to be done, it obviously needs some extensive
hand-holding.