node: use a linked-list for plugin-conf env vars
Avoid allocating space for 256 env vars in the conf struct
on the stack, when typically just a couple env vars are set.

Technically, using memory in a stack frame that exits for
putenv() was an error, because putenv() uses a reference
to the passed argument (unlike setenv() which copies them).